Rename boilerplate to kernel folder
