+++ src/globals.h
@@ -5,7 +5,7 @@
 
 #include "ux.h"
 
-#include "boilerplate/io.h"
+#include "kernel/io.h"
 #include "commands.h"
 #include "constants.h"
