+++ src/handler/handlers.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../boilerplate/dispatcher.h"
+#include "../kernel/dispatcher.h"
 
 void handler_get_extended_pubkey(dispatcher_context_t *dispatcher_context, uint8_t p2);
 void handler_get_master_fingerprint(dispatcher_context_t *dispatcher_context, uint8_t p2);
