+++ src/handler/lib/get_merkle_leaf_hash.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../boilerplate/dispatcher.h"
+#include "../../kernel/dispatcher.h"
 
 /**
  * TODO: docs
