+++ src/handler/sign_psbt/compare_wallet_script_at_path.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../boilerplate/dispatcher.h"
+#include "../../kernel/dispatcher.h"
 #include "../../common/merkle.h"
 #include "../../common/wallet.h"
