+++ src/handler/sign_psbt/update_hashes_with_map_value.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../boilerplate/dispatcher.h"
+#include "../../kernel/dispatcher.h"
 #include "../../common/merkle.h"
 
 /**
