+++ src/kernel/apdu_parser.c
