+++ src/kernel/apdu_parser.h
