+++ src/kernel/constants.h
