+++ src/kernel/io.c
