+++ src/kernel/offsets.h
