+++ src/ui/display_bagl.c
@@ -15,8 +15,8 @@
 #include "./display_utils.h"
 #include "../constants.h"
 #include "../globals.h"
-#include "../boilerplate/io.h"
-#include "../boilerplate/sw.h"
+#include "../kernel/io.h"
+#include "../kernel/sw.h"
 #include "../common/bip32.h"
 #include "../common/format.h"
 #include "../common/script.h"
